+++ publish/generate_docs_deploy_to_pypi.sh
@@ -1,0 +1,16 @@
+#!/bin/sh
+
+cd ../docs
+
+sphinx-apidoc -f -e -o source/ ../predictit
+
+cd ..
+
+python setup.py sdist bdist_wheel
+twine upload dist/*
+
+rm -r -f dist
+rm -r -f predictit.egg-info
+rm -r -f build
+
+ECHO Job done